file renamed, mass calculation moved to utils.py
