+++ README.md
@@ -2,7 +2,7 @@
 Automatic calculation of tardigrades biomass from microscopic images, using Deep Neural Networks.
 
 ### Build and run container
-To automatically build and run docker container run:
+To automatically build and run docker container, launch script:
 ```bash
 ./build_and_run_docker.sh
 ```
